# docs/openapi.yaml
openapi: 3.0.0
info:
  title: ViralHook AI API
  version: 1.0.0
  description: API for generating viral video hooks and scripts using AI.
servers:
  - url: https://viralhook.ai/api
    description: Production Server
  - url: http://localhost:3000/api
    description: Local Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
  schemas:
...
    HookRequest:
      type: object
      required:
        - script
      properties:
        script:
          type: string
          description: The video script to generate hooks for.
          minLength: 20
          maxLength: 2000
        stylePreference:
          type: string
        mode:
          type: string
          enum: [chaos, cinematic, shocking]
        targetModel:
          type: string
          enum: [kling, runway, luma, hailuo]
...
paths:
...
  /hooks:
    post:
      summary: Generate a visual hook for a script
      security:
        - bearerAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HookRequest'
      responses:
        '200':
          description: Successful hook generation
        '400':
          description: Invalid script or security violation
...